new: calculate x offset, replace buffer with displayed buffer (fixstr), fix: memory leak, shrink
The offset on x axis is created due to charactes being 1 byte but their displayed length is different. For example tab is one character but it is displayed as (usually)8. UTF-8 encodes characters outside ASCII as 2 or more bytes but displayed as one.

+++ utils/sizes.c
@@ -1,9 +1,9 @@
 #include "../headers/vars.hpp"
 
-char *itoa(int a)
+char *itoa(long a)
 {
 	static char b[12]; // static to keep in scope after function returns
-	sprintf(b, "%d", a);
+	sprintf(b, "%ld", a);
 	return b;
 }
